+++ test/unit/test_inference.py
@@ -20,7 +20,7 @@ import numpy as np
 import pytest
 import torch as pt
 
-import sockeye.beam_search_pt
+import sockeye.beam_search
 import sockeye.constants as C
 import sockeye.data_io
 import sockeye.inference_pt
